+++ tests/testthat/teardown.R
@@ -1,2 +1,3 @@
 options(old_opts)
-lg_rush$set_threshold(old_threshold_rush)
+lg$set_threshold(old_threshold)
+future::plan(old_plan)