+++ man-roxygen/param_message_log.R
@@ -1,0 +1,4 @@
+#' @param message_log (`character(1)`)\cr
+#' Path to the message log file.
+#' If `NULL`, no messages, warnings or errors are stored.
+