+++ man-roxygen/param_output_log.R
@@ -1,0 +1,4 @@
+#' @param output_log (`character(1)`)\cr
+#' Path to the output log file.
+#' If `NULL`, no output is stored.
+